Update bees_picamera.py
L 276: instead of hive.append(kp.pt), it should be hive.append(keypoints[i].pt)
Because in some case: for example when there is a bee in the entrance and a new bee is detected. It send (kp.pt) the position of the old bee and not the position (keypoints[i].pt) of the new bee. And as a side effect it increases the number of bees that went out.